config: move core config types into codex-config
