+++ codex-rs/config/src/lib.rs
@@ -11,6 +11,7 @@ mod project_root_markers;
 mod requirements_exec_policy;
 mod skills_config;
 mod state;
+pub mod types;
 
 pub const CONFIG_TOML_FILE: &str = "config.toml";
